+++ packages/riverpod/lib/src/providers/stream_notifier/family.dart
@@ -1,67 +1,9 @@
 part of '../stream_notifier.dart';
 
-/// {@macro riverpod.stream_notifier}
-///
-/// {@macro riverpod.async_notifier_provider_modifier}
-abstract class FamilyStreamNotifier<StateT, ArgT>
-    extends $StreamNotifier<StateT> {
-  /// {@template riverpod.notifier.family_arg}
-  /// The argument that was passed to this family.
-  ///
-  /// For example, when doing:
-  ///
-  /// ```dart
-  /// ref.watch(provider(0));
-  /// ```
-  ///
-  /// then [arg] will be `0`.
-  /// {@endtemplate}
-  late final ArgT arg = ref.$arg as ArgT;
-
-  /// {@macro riverpod.async_notifier.build}
-  @visibleForOverriding
-  Stream<StateT> build(ArgT arg);
-
-  @mustCallSuper
-  @override
-  void runBuild() {
-    final created = build(arg);
-    requireElement().handleValue(ref, created);
-  }
-}
-
-/// An implementation detail of Riverpod
-@publicInMisc
-final class FamilyStreamNotifierProvider< //
-        NotifierT extends FamilyStreamNotifier<ValueT, ArgT>,
-        ValueT,
-        ArgT> //
-    extends $StreamNotifierProvider<NotifierT, ValueT>
-    with LegacyProviderMixin<AsyncValue<ValueT>> {
-  /// An implementation detail of Riverpod
-  const FamilyStreamNotifierProvider._(
-    this._createNotifier, {
-    required super.name,
-    required super.dependencies,
-    required super.$allTransitiveDependencies,
-    required super.from,
-    required super.argument,
-    required super.isAutoDispose,
-    required super.retry,
-  });
-
-  final NotifierT Function() _createNotifier;
-
-  /// @nodoc
-  @internal
-  @override
-  NotifierT create() => _createNotifier();
-}
-
 /// The [Family] of [StreamNotifierProvider].
 @publicInMisc
 final class StreamNotifierProviderFamily< //
-        NotifierT extends FamilyStreamNotifier<ValueT, ArgT>,
+        NotifierT extends StreamNotifier<ValueT>,
         ValueT,
         ArgT> //
     extends ClassFamily< //
@@ -70,8 +12,8 @@ final class StreamNotifierProviderFamily< //
         ValueT,
         ArgT,
         Stream<ValueT>,
-        FamilyStreamNotifierProvider<NotifierT, ValueT, ArgT>> {
-  /// The [Family] of [FamilyStreamNotifierProvider].
+        StreamNotifierProvider<NotifierT, ValueT>> {
+  /// The [Family] of [StreamNotifierProvider].
   /// @nodoc
   @internal
   StreamNotifierProviderFamily.internal(
@@ -81,7 +23,7 @@ final class StreamNotifierProviderFamily< //
     super.isAutoDispose = false,
     super.retry,
   }) : super(
-          providerFactory: FamilyStreamNotifierProvider._,
+          providerFactory: StreamNotifierProvider.internal,
           $allTransitiveDependencies:
               computeAllTransitiveDependencies(dependencies),
         );