+++ packages/riverpod/lib/src/providers/stream_notifier/orphan.dart
@@ -12,8 +12,6 @@ part of '../stream_notifier.dart';
 /// [AsyncNotifier].
 ///
 /// This can be considered as a [StreamProvider] that can mutate its value over time.
-/// When using `family`, your notifier type changes. Instead of extending
-/// [StreamNotifier], you should extend [FamilyStreamNotifier].
 /// {@endtemplate}
 /// {@category Notifiers}
 abstract class StreamNotifier<ValueT> extends $StreamNotifier<ValueT> {
@@ -40,11 +38,6 @@ abstract class StreamNotifier<ValueT> extends $StreamNotifier<ValueT> {
 /// Instead the ref (and argument) are directly accessible in the associated
 /// [StreamNotifier].
 /// {@endtemplate}
-///
-/// {@template riverpod.async_notifier_provider_modifier}
-/// When using your notifier type changes.
-/// Instead of extending [StreamNotifier], you should extend [FamilyStreamNotifier].
-/// {@endtemplate}
 /// {@category Providers}
 final class StreamNotifierProvider< //
         NotifierT extends StreamNotifier<ValueT>,
@@ -52,8 +45,6 @@ final class StreamNotifierProvider< //
     extends $StreamNotifierProvider<NotifierT, ValueT>
     with LegacyProviderMixin<AsyncValue<ValueT>> {
   /// {@macro riverpod.stream_notifier_provider}
-  ///
-  /// {@macro riverpod.async_notifier_provider_modifier}
   StreamNotifierProvider(
     this._createNotifier, {
     super.name,