+++ packages/riverpod/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix provider rebuild order issue.
 - Fix "Tried to refresh x multiple times in the same frame" incorrectly triggering.
+- Removed `FamilyNotifier` and variants, in favour of `Notifier`.
 
 ## 3.0.0-dev.17 - 2025-08-01